feat: new rush game started from scrath
